move: Update types folder structure
