+++ src/types/ErrorTypes.ts
@@ -1,0 +1,4 @@
+export type ErrorAPIResponse = {
+  message: string;
+  issues?: string[];
+};
