+++ src/types/PopulationTypes.ts
@@ -1,21 +1,4 @@
-export type ErrorAPIResponse = {
-  message: string;
-  issues?: string[];
-};
-
-type PrefectureBase = {
-  prefCode: number;
-  prefName: string;
-};
-
-export type PrefecturesAPIResponse = {
-  message: string | null;
-  result: PrefectureBase[];
-};
-
-export type PrefectureWithCheck = PrefectureBase & {
-  isChecked: boolean;
-};
+import type { PrefectureBase } from './PrefectureTypes';
 
 type PopulationDataItem = {
   year: number;
